Refactoring code and add comments
