+++ src/imageLoader/wadors/metaDataManager.js
@@ -1,66 +1,7 @@
 import imageIdToURI from '../imageIdToURI.js';
+import combineFrameInstance from "./combineFrameInstance"
 
 let metadataByImageURI = [];
-
-function getValue(tag, justElement = true) {
-  if (tag) {
-    if (tag.Value) {
-      if (tag.Value[0] && justElement) {
-        return tag.Value[0];
-      }
-
-      return tag.Value;
-    }
-  }
-
-  return tag;
-}
-
-function combineFrameInstance(frame, instance) {
-  let {
-    52009230: PerFrameFunctionalGroupsSequence,
-    52009229: SharedFunctionalGroupsSequence,
-    '00280008': NumberOfFrames,
-    // eslint-disable-next-line prefer-const
-    ...rest
-  } = instance;
-
-  PerFrameFunctionalGroupsSequence = getValue(
-    PerFrameFunctionalGroupsSequence,
-    false
-  );
-  SharedFunctionalGroupsSequence = getValue(
-    SharedFunctionalGroupsSequence,
-    false
-  );
-  NumberOfFrames = getValue(NumberOfFrames);
-
-  if (PerFrameFunctionalGroupsSequence || NumberOfFrames > 1) {
-    const frameNumber = Number.parseInt(frame || 1, 10);
-    const shared = (
-      SharedFunctionalGroupsSequence
-        ? Object.values(SharedFunctionalGroupsSequence[0])
-        : []
-    )
-      .map((it) => it[0])
-      .filter((it) => it !== undefined && typeof it === 'object');
-    const perFrame = (
-      PerFrameFunctionalGroupsSequence
-        ? Object.values(PerFrameFunctionalGroupsSequence[frameNumber - 1])
-        : []
-    )
-      .map((it) => it.Value[0])
-      .filter((it) => it !== undefined && typeof it === 'object');
-
-    return Object.assign(
-      rest,
-      ...Object.values(shared),
-      ...Object.values(perFrame)
-    );
-  }
-
-  return instance;
-}
 
 function add(imageId, metadata) {
   const imageURI = imageIdToURI(imageId);
@@ -71,22 +12,44 @@ function add(imageId, metadata) {
 function get(imageId) {
   const imageURI = imageIdToURI(imageId);
 
+  // multiframes images will have only one imageid returned by the dicomweb
+  // client and registered in metadataByImageURI for all the n frames. If an
+  // iamgeid does not have metadata, or it does not have at all, or the imageid
+  // belongs to a frame, not registered in metadataByImageURI
   let metadata = metadataByImageURI[imageURI];
-
   if (!metadata) {
-    // in this case try see if the imageId corresponds to multiframe
+    // in this case it might indicate a multiframe imageid
+    const lastSlashIdx = imageURI.indexOf('/frames/') + 8;
     const imageIdFrameless = imageURI.slice(
       0,
-      imageURI.indexOf('/frames/') + 8
+      lastSlashIdx
     );
+    // get all text after the last slash and convert to a frame number
     const frame = parseInt(
-      imageURI.slice(imageURI.indexOf('/frames/') + 9),
+      imageURI.slice(lastSlashIdx + 1),
       10
     );
-
+    // retrieving the frame 1 that contains all multiframe information
     metadata = metadataByImageURI[`${imageIdFrameless}1`];
     if (metadata) {
+      // create the metadata frame
       metadata = combineFrameInstance(frame, metadata);
+    }
+  }
+  else
+  {
+    // if the metadata exists, test if this is the multiframe imageid
+    if (metadata['00280008'] && metadata['00280008'] > 1)
+    {
+      const lastSlashIdx = imageURI.indexOf('/frames/') + 8;
+      const frame = parseInt(
+        imageURI.slice(lastSlashIdx + 1),
+        10
+      );
+      // if is multiframe imageid, return only the frame 1 metadata
+      if (frame===1) {
+        metadata = combineFrameInstance(frame, metadata);
+      }
     }
   }
